move BoardTile to another file
